Fixes after review
