+++ client/src/utils/getRandomBetween.js
@@ -1,0 +1,7 @@
+// Get random number between minand max
+const getRandomBetween = (min, max) => {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
+module.exports = getRandomBetween;
